refactor: split game models and add Shikona tests
