+++ game/models/__init__.py
@@ -1,0 +1,6 @@
+"""Database models for the game app."""
+
+from .shikona import Shikona
+from .shusshin import Shusshin
+
+__all__ = ["Shusshin", "Shikona"]
