+++ game/models/shusshin.py
@@ -1,4 +1,4 @@
-"""Database models for the game app."""
+"""Origin model for sumo wrestlers."""
 
 from django.db import models
 from django.db.models import Q
@@ -48,23 +48,3 @@ class Shusshin(models.Model):
         if self.country_code == Country.JP:
             return JPPrefecture(self.jp_prefecture).label
         return Country(self.country_code).label
-
-
-class Shikona(models.Model):
-    """Ring name for a sumo wrestler."""
-
-    name = models.CharField(max_length=8, unique=True)
-    transliteration = models.CharField(max_length=32, unique=True)
-    interpretation = models.CharField(max_length=64)
-
-    class Meta:
-        """Model metadata."""
-
-        ordering = ["transliteration"]
-        verbose_name = "Shikona"
-        verbose_name_plural = "Shikona"
-        unique_together = ("name", "transliteration")
-
-    def __str__(self) -> str:
-        """Return the ring name."""
-        return f"{self.transliteration} ({self.name})"
